requirements.txt
appdirs==1.4.4
asgiref==3.3.1
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.4
click==7.1.2
coverage==5.4
decorator==4.4.2
distlib==0.3.1
Django==3.1.4
djangorestframework==3.12.2
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
flipflop==1.0
guess-language==0.2
idna==3.1
is-disposable-email==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysqlclient==2.0.3
pbr==5.5.1
pycparser==2.20
pymongo==3.11.3
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
Tempita==0.5.2
virtualenv==20.2.2
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.3
wtforms-validators==1.0.0
